+++ lib/src/services/widget_parser_service.dart
@@ -3,6 +3,7 @@ import 'package:flutter/widgets.dart';
 import '../adapters/card_adapter.dart';
 import '../adapters/common_widget_adapter.dart';
 import '../adapters/list_tile_adapter.dart';
+import '../adapters/stateless_widget_adapter.dart';
 import '../builders/widget_transformer.dart';
 import '../models/shimmer_node.dart';
 import '../transformers/container_transformer.dart';
@@ -25,6 +26,7 @@ class WidgetParserService {
     CardAdapter(),
     ContainerTransformer(),
     LayoutTransformer(),
+    StatelessWidgetAdapter(),
     CommonWidgetAdapter(),
   ];